requirements/test_requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements/test_requirements.txt requirements/test_requirements.in
#
appdirs==1.4.3            # via black
atomicwrites==1.3.0       # via pytest
attrs==19.3.0             # via black, pytest
black==19.10b0            # via -r requirements/test_requirements.in (line 9)
click==7.0                # via black, pip-tools
coverage==5.0.3           # via pytest-cov
entrypoints==0.3          # via flake8
filelock==3.0.12          # via pytest-mypy
flake8==3.7.9             # via -r requirements/test_requirements.in (line 2)
importlib-metadata==1.5.0  # via pluggy, pytest
mccabe==0.6.1             # via flake8
mock==3.0.5               # via -r requirements/test_requirements.in (line 10)
more-itertools==8.2.0     # via pytest
mypy-extensions==0.4.3    # via mypy
mypy==0.761               # via -r requirements/test_requirements.in (line 3), pytest-mypy
packaging==20.1           # via pytest
pathspec==0.7.0           # via black
pip-tools==4.5.0          # via -r requirements/test_requirements.in (line 8)
pluggy==0.13.1            # via pytest
py==1.8.1                 # via pytest
pycodestyle==2.5.0        # via flake8
pyflakes==2.1.1           # via flake8, pytest-flakes
pyparsing==2.4.6          # via packaging
pytest-cov==2.8.1         # via -r requirements/test_requirements.in (line 5)
pytest-flakes==4.0.0      # via -r requirements/test_requirements.in (line 7)
pytest-mypy==0.5.0        # via -r requirements/test_requirements.in (line 6)
pytest==4.6.9             # via -r requirements/test_requirements.in (line 4), pytest-cov, pytest-flakes, pytest-mypy
pyyaml==5.3               # via -r requirements/test_requirements.in (line 11)
regex==2020.2.20          # via black
six==1.14.0               # via -c requirements/requirements.txt (line 20), mock, packaging, pip-tools, pytest
toml==0.10.0              # via black
typed-ast==1.4.1          # via black, mypy
typing-extensions==3.7.4.1  # via mypy
wcwidth==0.1.8            # via pytest
zipp==3.0.0               # via importlib-metadata
